Adding 3D files
In this commit, the 3D files needed to print the floating clock were added. The files include the CAD model of the clock, components for the floating sphere, and other details needed to 3D print the design. This addition will allow users to view and modify the design of the clock according to their preferences, thus contributing to the customization of the design.