Addition of Arduino code
In this commit, the source code for Arduino needed to operate the floating clock was included. The code handles the servomotor control that controls the position of the floating ball, thus enabling the device to function as intended. This addition makes the repository complete, providing all the files essential to the project's implementation and making it easy for developers to understand and modify the code to suit their needs.